% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{apply_adapters}
\alias{apply_adapters}
\title{Apply adapter functions}
\usage{
apply_adapters(df, adapters)
}
\arguments{
\item{df}{The dataframe to process}

\item{adapters}{A list of lists, containing keys 'detects' (function) and 'apply' (function)}
}
\description{
Apply adapter functions
}
\details{
Adapter functions perform specific actions for some data types, e.g. parse session start and end time in GGIR data
}
\keyword{internal}
