% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{apply_rules}
\alias{apply_rules}
\title{Apply rules}
\usage{
apply_rules(df, rules, max_iter = 10)
}
\arguments{
\item{df}{The dataframe to process}

\item{rules}{a list of lists containing keys 'requires' (str, column name), 'produces' (str, output column name), and 'apply' (function)}
}
\description{
Apply rules
}
\details{
Rules create missing columns based on those that are present in the data.
The function iterates over all rules to make changes until all possible changes have been made.
}
\keyword{internal}
