% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{chronotype}
\alias{chronotype}
\title{Calculate the Chronotype}
\usage{
chronotype(sessions)
}
\arguments{
\item{sessions}{The sessions data frame}
}
\value{
The Chronotype value in hours
}
\description{
This function calculates the Chronotype metric based on the mid-sleep time
If sleep duration on free days is greater than on workdays, it applies a correction
as described in Roenneberg et al. (2019).
}
\details{
This function uses columns:
\itemize{
\item \code{time_at_midsleep}
\item \code{sleep_period}
\item \code{is_workday}
}
}
\examples{
chronotype(example_sessions)
}
\seealso{
Other sleep metrics:
\code{\link[=composite_phase_deviation]{composite_phase_deviation()}},
\code{\link[=interdaily_stability]{interdaily_stability()}},
\code{\link[=sleep_regularity_index]{sleep_regularity_index()}},
\code{\link[=social_jet_lag]{social_jet_lag()}}
}
\concept{sleep metrics}
