% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{circ_time_diff}
\alias{circ_time_diff}
\title{Compute the signed time difference from t1 to t2}
\usage{
circ_time_diff(t1, t2, unit = "hour")
}
\arguments{
\item{t1}{First time (character, POSIXct, or numeric hour)}

\item{t2}{Second time (character, POSIXct, or numeric hour)}

\item{unit}{The unit of time. Can be "second", "minute", or "hour" (default)}
}
\value{
The signed time difference in the specified unit
}
\description{
Returns the circular time difference between t1 and t2
}
\examples{
circ_time_diff("07:00", "22:00") # 9
circ_time_diff("22:00", "07:00") # -9
circ_time_diff("07:00", "22:00", unit = "minute") # 540
}
\concept{time_processing}
