% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_accessors.R
\name{colnames_to_canonical}
\alias{colnames_to_canonical}
\title{Rename sessions columns to canonical names}
\usage{
colnames_to_canonical(df)
}
\arguments{
\item{df}{The input dataframe (sessions or epochs) to rename columns on}
}
\value{
The updated dataframe with missing canonical columns created
}
\description{
Rename sessions columns to canonical names
}
\keyword{internal}
