% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_key.R
\name{.epochs_col_presets}
\alias{.epochs_col_presets}
\title{Preset values for epoch column names}
\usage{
.epochs_col_presets
}
\value{
A list of canonical epoch column names, where each value is an array of possible matching columns in the data (strings)
}
\description{
Preset values for epoch column names
}
\keyword{internal}
