% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_key.R
\name{.epochs_long}
\alias{.epochs_long}
\title{Long names for epochs columns, for display in app}
\usage{
.epochs_long
}
\value{
A list matching canonical epochs column names to their long name
}
\description{
Long names for epochs columns, for display in app
}
\keyword{internal}
