% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_key.R
\name{.epochs_parser}
\alias{.epochs_parser}
\title{Parsing functions to use for each epochs column type}
\usage{
.epochs_parser()
}
\description{
Parsing functions to use for each epochs column type
}
\keyword{internal}
