% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_key.R
\name{.sessions_col_presets}
\alias{.sessions_col_presets}
\title{Preset values for session column names}
\usage{
.sessions_col_presets
}
\value{
A list of canonical session column names, where each value is an array of possible matching columns in the data (strings)
}
\description{
Preset values for session column names
}
\keyword{internal}
