% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_key.R
\name{.sessions_long}
\alias{.sessions_long}
\title{Long names for sessions columns, for display in app}
\usage{
.sessions_long
}
\value{
A list matching canonical sessions column names to their long name
}
\description{
Long names for sessions columns, for display in app
}
\keyword{internal}
