% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_key.R
\name{.sessions_parsers}
\alias{.sessions_parsers}
\title{Parsing functions to use for each sessions column type}
\usage{
.sessions_parsers()
}
\description{
Parsing functions to use for each sessions column type
}
\keyword{internal}
