% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_by_age_range}
\alias{filter_by_age_range}
\title{Filter sessions by age range}
\usage{
filter_by_age_range(
  sessions,
  min_age = NULL,
  max_age = NULL,
  return_mask = FALSE
)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{min_age}{The minimum age of the subjects (inclusive)}

\item{max_age}{The maximum age of the subjects (inclusive)}

\item{return_mask}{If TRUE, returns a logical vector indicating which sessions belong to subjects within the specified age range}
}
\value{
The sessions dataframe with only the sessions that belong to subjects within the specified age range,
or a logical vector if \code{return_mask} is TRUE
}
\description{
Filter sessions by age range
}
\details{
This function uses columns:
\itemize{
\item \code{birth_year}
\item \code{night}
}
}
\examples{
filtered_sessions <- filter_by_age_range(example_sessions_v1, min_age = 11, max_age = 18)
}
\seealso{
Other filtering:
\code{\link[=filter_by_night_range]{filter_by_night_range()}},
\code{\link[=filter_by_sex]{filter_by_sex()}},
\code{\link[=filter_epochs_from_sessions]{filter_epochs_from_sessions()}},
\code{\link[=keep_longest]{keep_longest()}},
\code{\link[=remove_sessions_no_sleep]{remove_sessions_no_sleep()}},
\code{\link[=select_devices]{select_devices()}},
\code{\link[=select_subjects]{select_subjects()}},
\code{\link[=set_min_sleep_period]{set_min_sleep_period()}},
\code{\link[=set_min_time_in_bed]{set_min_time_in_bed()}},
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}},
\code{\link[=set_session_start_time_range]{set_session_start_time_range()}}
}
\concept{filtering}
