% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_by_sex}
\alias{filter_by_sex}
\title{Filter by sex}
\usage{
filter_by_sex(sessions, sex, return_mask = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{sex}{The sex to filter for (M, F, or NULL for both)}

\item{return_mask}{If TRUE, returns a logical vector indicating which sessions belong to the specified sex}
}
\value{
The sessions dataframe with only the sessions that belong to the specified sex, or a logical vector if \code{return_mask} is TRUE
}
\description{
Filter by sex
}
\details{
This function uses columns:
\itemize{
\item \code{sex}
}
}
\examples{
filtered_sessions <- filter_by_sex(example_sessions_v1, "M")
}
\seealso{
Other filtering:
\code{\link[=filter_by_age_range]{filter_by_age_range()}},
\code{\link[=filter_by_night_range]{filter_by_night_range()}},
\code{\link[=filter_epochs_from_sessions]{filter_epochs_from_sessions()}},
\code{\link[=remove_sessions_no_sleep]{remove_sessions_no_sleep()}},
\code{\link[=select_devices]{select_devices()}},
\code{\link[=select_subjects]{select_subjects()}},
\code{\link[=set_min_sleep_period]{set_min_sleep_period()}},
\code{\link[=set_min_time_in_bed]{set_min_time_in_bed()}},
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}},
\code{\link[=set_session_start_time_range]{set_session_start_time_range()}}
}
\concept{filtering}
