% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_epochs_from_sessions}
\alias{filter_epochs_from_sessions}
\title{Filter epochs based on session IDs}
\usage{
filter_epochs_from_sessions(epochs, sessions, return_mask = FALSE)
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{sessions}{The sessions dataframe}

\item{return_mask}{If TRUE, returns a logical vector indicating which epochs belong to the specified sessions}
}
\value{
The epochs dataframe with only the epochs that belong to the specified sessions, or a logical vector if \code{return_mask} is TRUE
}
\description{
Filter epochs based on session IDs
}
\details{
This function uses sessions columns:
\itemize{
\item \code{id}
And epoch columns:
\item \code{session_id}
}
}
\examples{
# Apply filtering to sessions to keep specific nights, and filter epochs accordingly
filtered_sessions <- filter_by_night_range(example_sessions, "2025-04-07", "2025-04-10")
filtered_epochs <- filter_epochs_from_sessions(example_epochs, filtered_sessions)

}
\seealso{
\code{\link[=filter_by_night_range]{filter_by_night_range()}} to filter sessions by night range.

Other filtering:
\code{\link[=filter_by_age_range]{filter_by_age_range()}},
\code{\link[=filter_by_night_range]{filter_by_night_range()}},
\code{\link[=filter_by_sex]{filter_by_sex()}},
\code{\link[=keep_longest]{keep_longest()}},
\code{\link[=remove_sessions_no_sleep]{remove_sessions_no_sleep()}},
\code{\link[=select_devices]{select_devices()}},
\code{\link[=select_subjects]{select_subjects()}},
\code{\link[=set_min_sleep_period]{set_min_sleep_period()}},
\code{\link[=set_min_time_in_bed]{set_min_time_in_bed()}},
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}},
\code{\link[=set_session_start_time_range]{set_session_start_time_range()}}
}
\concept{filtering}
