% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_accessors.R
\name{get_colnames}
\alias{get_colnames}
\title{Get standard column names for a data frame (sessions or epochs)}
\usage{
get_colnames(df)
}
\arguments{
\item{df}{A data frame containing either sessions or epochs data}
}
\value{
A named list of column names
}
\description{
Dispatch to the appropriate function based on the dataframe type attribute (expecting "sessions" or "epochs").
}
\examples{
get_colnames(example_sessions)
}
\seealso{
Other internal:
\code{\link[=check_suggests]{check_suggests()}},
\code{\link[=get_col]{get_col()}},
\code{\link[=get_epoch_colnames]{get_epoch_colnames()}},
\code{\link[=get_session_colnames]{get_session_colnames()}},
\code{\link[=get_time_per_day]{get_time_per_day()}},
\code{\link[=is_iso8601_datetime]{is_iso8601_datetime()}},
\code{\link[=set_colnames]{set_colnames()}},
\code{\link[=set_data_type]{set_data_type()}}
}
\concept{internal}
