% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tables.R
\name{get_epochs_summary}
\alias{get_epochs_summary}
\title{Summarise epoch information}
\usage{
get_epochs_summary(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
A single-row dataframe summarising epoch information
}
\description{
Display the number of sessions in the epoch data, as well as the start and end dates of the epoch data
}
\details{
This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{session_id}
}
}
\examples{
get_epochs_summary(example_epochs)
}
\seealso{
\code{\link[=get_sessions_summary]{get_sessions_summary()}} to summarise session information.

Other data tables:
\code{\link[=get_non_complying_sessions]{get_non_complying_sessions()}},
\code{\link[=get_removed_sessions]{get_removed_sessions()}},
\code{\link[=get_sessions_summary]{get_sessions_summary()}}
}
\concept{data tables}
