% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_accessors.R
\name{get_session_colnames}
\alias{get_session_colnames}
\title{Get a list of standard column names for sessions data}
\usage{
get_session_colnames(sessions)
}
\arguments{
\item{sessions}{A data frame containing sessions data}
}
\value{
A named list of column names
}
\description{
This function retrieves the standard column names for sessions data.
If the column names are not set as an attribute, it attempts to infer them
from a list of common column names.
}
\examples{
get_session_colnames(example_sessions)
}
\seealso{
Other internal:
\code{\link[=check_suggests]{check_suggests()}},
\code{\link[=get_col]{get_col()}},
\code{\link[=get_colnames]{get_colnames()}},
\code{\link[=get_epoch_colnames]{get_epoch_colnames()}},
\code{\link[=get_time_per_day]{get_time_per_day()}},
\code{\link[=is_iso8601_datetime]{is_iso8601_datetime()}},
\code{\link[=set_colnames]{set_colnames()}},
\code{\link[=set_data_type]{set_data_type()}}
}
\concept{internal}
