% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{get_time_per_day}
\alias{get_time_per_day}
\title{Calculate the number of time units in a day}
\usage{
get_time_per_day(unit = "second")
}
\arguments{
\item{unit}{The unit of time. Can be "second", "minute", or "hour". Default is "second".}
}
\value{
The number of time units in a day (numeric)
}
\description{
This function returns the number of specified time units (seconds, minutes or hours) in a day.
}
\examples{
get_time_per_day("hour")
}
\seealso{
Other internal:
\code{\link[=check_suggests]{check_suggests()}},
\code{\link[=get_col]{get_col()}},
\code{\link[=get_colnames]{get_colnames()}},
\code{\link[=get_epoch_colnames]{get_epoch_colnames()}},
\code{\link[=get_session_colnames]{get_session_colnames()}},
\code{\link[=is_iso8601_datetime]{is_iso8601_datetime()}},
\code{\link[=set_colnames]{set_colnames()}},
\code{\link[=set_data_type]{set_data_type()}}
}
\concept{internal}
