% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{group_epochs_by_night}
\alias{group_epochs_by_night}
\title{Create a grouping by night for epoch data}
\usage{
group_epochs_by_night(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
The epochs dataframe with the \code{night} column added
}
\description{
The function creates a new column \code{night} that groups the epochs by night.
Timepoints before 12 PM are considered part of the previous night.
}
\details{
This function uses columns:
\itemize{
\item \code{timestamp}
}
}
\examples{
epochs <- group_epochs_by_night(example_epochs)
}
\seealso{
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}} to group session data by night.

Other time processing:
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}},
\code{\link[=max_time]{max_time()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_date]{parse_date()}},
\code{\link[=parse_time]{parse_time()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=shift_times_by_12h]{shift_times_by_12h()}},
\code{\link[=time_diff]{time_diff()}},
\code{\link[=time_to_hours]{time_to_hours()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
