% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{group_sessions_by_night}
\alias{group_sessions_by_night}
\title{Create a grouping by night for session data}
\usage{
group_sessions_by_night(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe}
}
\value{
The sessions dataframe with the \code{night} column added
}
\description{
The function creates a new column \code{night} that groups the sessions by night depending on their start time.
Sessions that start before 12 PM are considered part of the previous night.
}
\details{
This function uses columns:
\itemize{
\item \code{session_start}
}
}
\examples{
sessions <- group_sessions_by_night(example_sessions)
}
\seealso{
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}} to group epoch data by night.

Other time processing:
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}},
\code{\link[=max_time]{max_time()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_date]{parse_date()}},
\code{\link[=parse_time]{parse_time()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=shift_times_by_12h]{shift_times_by_12h()}},
\code{\link[=time_diff]{time_diff()}},
\code{\link[=time_to_hours]{time_to_hours()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
