% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{interdaily_stability}
\alias{interdaily_stability}
\title{Calculate Interdaily Stability (IS)}
\usage{
interdaily_stability(epochs)
}
\arguments{
\item{epochs}{The epochs data frame}
}
\value{
The Interdaily Stability (IS) value
}
\description{
This function calculates the Interdaily Stability (IS) metric from a binary awake/asleep variable
}
\details{
This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{is_asleep}
}
}
\examples{
interdaily_stability(example_epochs)
}
\seealso{
Other sleep metrics:
\code{\link[=chronotype]{chronotype()}},
\code{\link[=composite_phase_deviation]{composite_phase_deviation()}},
\code{\link[=sleep_regularity_index]{sleep_regularity_index()}},
\code{\link[=social_jet_lag]{social_jet_lag()}}
}
\concept{sleep metrics}
