% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{is_iso8601_datetime}
\alias{is_iso8601_datetime}
\title{Check if a column contains ISO 8601 datetime strings}
\usage{
is_iso8601_datetime(x)
}
\arguments{
\item{x}{A vector of character strings}
}
\value{
TRUE if all non-missing values are valid ISO 8601 datetime strings, FALSE otherwise
}
\description{
This function checks if all non-missing values in a column are valid ISO 8601 datetime strings.
}
\examples{
is_iso8601_datetime(c("2026-01-01T14:10"))
}
\seealso{
Other internal:
\code{\link[=check_suggests]{check_suggests()}},
\code{\link[=get_col]{get_col()}},
\code{\link[=get_colnames]{get_colnames()}},
\code{\link[=get_epoch_colnames]{get_epoch_colnames()}},
\code{\link[=get_session_colnames]{get_session_colnames()}},
\code{\link[=get_time_per_day]{get_time_per_day()}},
\code{\link[=set_colnames]{set_colnames()}},
\code{\link[=set_data_type]{set_data_type()}}
}
\concept{internal}
