% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{keep_longest}
\alias{keep_longest}
\title{Remove duplicate sessions by keeping only the longest}
\usage{
keep_longest(sessions)
}
\arguments{
\item{sessions}{The sessions dataframe to process}
}
\value{
The sessions dataframe with duplicate sessions removed
}
\description{
Remove duplicate sessions by keeping only the longest
}
\details{
This function uses columns:
\itemize{
\item \code{night}
\item \code{sleep_period}
}
}
\examples{
keep_longest(example_sessions)
}
\seealso{
Other filtering:
\code{\link[=filter_by_age_range]{filter_by_age_range()}},
\code{\link[=filter_by_night_range]{filter_by_night_range()}},
\code{\link[=filter_by_sex]{filter_by_sex()}},
\code{\link[=filter_epochs_from_sessions]{filter_epochs_from_sessions()}},
\code{\link[=remove_sessions_no_sleep]{remove_sessions_no_sleep()}},
\code{\link[=select_devices]{select_devices()}},
\code{\link[=select_subjects]{select_subjects()}},
\code{\link[=set_min_sleep_period]{set_min_sleep_period()}},
\code{\link[=set_min_time_in_bed]{set_min_time_in_bed()}},
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}},
\code{\link[=set_session_start_time_range]{set_session_start_time_range()}}
}
\concept{filtering}
