% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_sessions}
\alias{load_sessions}
\title{Load session data}
\usage{
load_sessions(sessions_file)
}
\arguments{
\item{sessions_file}{The path to the sessions file}
}
\value{
A dataframe containing the session data
}
\description{
Load session data
}
\details{
The function loads the session data from a file
Supported formats: CSV, Excel, EDF.
}
\examples{
tf <- tempfile(fileext = ".csv")
utils::write.csv(data.frame(id = 1), tf, row.names = FALSE)
load_sessions(tf)
}
\seealso{
Other data loading:
\code{\link[=load_batch]{load_batch()}},
\code{\link[=load_epochs]{load_epochs()}},
\code{\link[=read_edf_epochs]{read_edf_epochs()}},
\code{\link[=read_edf_sessions]{read_edf_sessions()}}
}
\concept{data loading}
