% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{max_time}
\alias{max_time}
\title{Calculate the maximum time from 12pm to 12pm}
\usage{
max_time(time_vector)
}
\arguments{
\item{time_vector}{A vector of time strings in the format "YYYY-MM-DD HH:MM:SS".}
}
\value{
A string representing the maximum time in the format "HH:MM".
}
\description{
This function calculates the maximum time from a vector of time strings in the format "YYYY-MM-DD HH:MM:SS".
It considers a time window from 12pm to 12pm the next day, so 11:00 is considered later than 13:00.
}
\examples{
max_time(c("2025-04-08 23:00:00", "2025-04-09 01:00:00", "2025-04-09 02:30:00"))
}
\seealso{
\code{\link[=min_time]{min_time()}} to calculate the minimum time in the same format.

Other time processing:
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}},
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_date]{parse_date()}},
\code{\link[=parse_time]{parse_time()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=shift_times_by_12h]{shift_times_by_12h()}},
\code{\link[=time_diff]{time_diff()}},
\code{\link[=time_to_hours]{time_to_hours()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
