% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocturn.R
\name{nocturn}
\alias{nocturn}
\title{nocturn app}
\usage{
nocturn()
}
\value{
No return value, called for side-effects
}
\description{
This function launches the nocturn app, a Shiny application for visualizing and analyzing sleep data.
}
\examples{
if(interactive()){nocturn()}
}
