% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{parse_date}
\alias{parse_date}
\title{Parse a vector of date strings into Date objects}
\usage{
parse_date(date_vector)
}
\arguments{
\item{date_vector}{A vector of date strings}
}
\value{
A vector of Date objects
}
\description{
This function parses a vector of date strings into Date objects.
All formats containing year, month, and day information are supported.
}
\examples{
parse_date("2026-01-02T14:33:09")
}
\seealso{
Other time processing:
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}},
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}},
\code{\link[=max_time]{max_time()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_time]{parse_time()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=shift_times_by_12h]{shift_times_by_12h()}},
\code{\link[=time_diff]{time_diff()}},
\code{\link[=time_to_hours]{time_to_hours()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
