% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{parse_time}
\alias{parse_time}
\title{Parse a vector of time strings into POSIXct objects}
\usage{
parse_time(time_vector)
}
\arguments{
\item{time_vector}{A vector of time strings}
}
\value{
A vector of POSIXct objects
}
\description{
This function parses a vector of time strings into POSIXct objects.
Supported formats include "YYYY-MM-DD HH:MM:SS", "YYYY-MM-DD HH:MM", "HH:MM:SS", and "HH:MM".
Timezone information is ignored.
}
\examples{
parse_time("2026-01-01T14:34:09")
}
\seealso{
Other time processing:
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}},
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}},
\code{\link[=max_time]{max_time()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_date]{parse_date()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=shift_times_by_12h]{shift_times_by_12h()}},
\code{\link[=time_diff]{time_diff()}},
\code{\link[=time_to_hours]{time_to_hours()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
