% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\title{Create a Bland-Altman plot}
\usage{
plot_bland_altman(sessions1, sessions2, variable)
}
\arguments{
\item{sessions1}{The first sessions dataframe to be compared}

\item{sessions2}{The second sessions dataframe to be compared}

\item{variable}{The name of the variable to compare. Can be a single value for both dataframes, or an array of two values, one for each dataframe}
}
\value{
a ggplot object showing the Bland-Altman plot
}
\description{
Create a Bland-Altman plot
}
\details{
This function uses columns:
\itemize{
\item \code{night}
\item \code{sleep_period}
}
}
\examples{
  sessions1 <- data.frame(
   time_at_wakeup = c("2025-03-03T11:24:11.524000+00:00",
                      "2025-03-03T14:44:11.524000+00:00",
                      "2025-03-03T18:24:11.524000+00:00",
                      "2025-03-03T00:24:11.524000+00:00",
                      "2025-03-03T23:24:11.524000+00:00"),
   night = c("2025-03-02", "2025-03-03", "2025-03-03", "2025-03-03", "2025-03-03"),
   sleep_period = c(1000, 1000, 10, 2000, 1500)
 )

sessions2 <- data.frame(
   time_at_wakeup = c("2025-03-03T11:14:11.524000+00:00",
                      "2025-03-03T14:24:11.524000+00:00",
                      "2025-03-03T18:00:11.524000+00:00",
                      "2025-03-03T00:22:11.524000+00:00",
                      "2025-03-03T23:12:11.524000+00:00"),
   night = c("2025-03-02", "2025-03-03", "2025-03-03", "2025-03-03", "2025-03-03"),
   sleep_period = c(1000, 1000, 10000, 200, 1500)
 )
plot_bland_altman(sessions1, sessions2, "time_at_wakeup")
}
\concept{comparison plot}
