% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypnogram.R
\name{plot_hypnogram}
\alias{plot_hypnogram}
\title{Plot Hypnogram}
\usage{
plot_hypnogram(epochs)
}
\arguments{
\item{epochs}{The epochs dataframe}
}
\value{
A ggplot object showing the hypnogram as bars
}
\description{
Plot Hypnogram
}
\details{
This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{sleep_stage}
}
}
\examples{
plot_hypnogram(example_epochs)
}
\seealso{
Other plot epochs:
\code{\link[=plot_sleep_spiral]{plot_sleep_spiral()}},
\code{\link[=plot_timeseries]{plot_timeseries()}}
}
\concept{plot epochs}
