% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_spiral.R
\name{plot_sleep_spiral}
\alias{plot_sleep_spiral}
\title{Plot Sleep Spiral}
\usage{
plot_sleep_spiral(epochs, color_by = "default")
}
\arguments{
\item{epochs}{The epochs dataframe}

\item{color_by}{The variable to color the spiral by. Can be "default" or any other column name in the epochs dataframe.}
}
\value{
A ggplot object showing the sleep spiral
}
\description{
Plot Sleep Spiral
}
\details{
This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{is_asleep}
}
}
\examples{
plot_sleep_spiral(example_epochs)
}
\seealso{
Other plot epochs:
\code{\link[=plot_hypnogram]{plot_hypnogram()}},
\code{\link[=plot_timeseries]{plot_timeseries()}}
}
\concept{plot epochs}
