% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_reading.R
\name{read_edf_epochs}
\alias{read_edf_epochs}
\title{Read EDF Epochs}
\usage{
read_edf_epochs(file)
}
\arguments{
\item{file}{The path to the EDF file}
}
\value{
A dataframe containing the epoch data extracted from the EDF file signals
}
\description{
Read EDF Epochs
}
\details{
The function reads the signals of the EDF file to extract epoch information. It must contain
a column for timestamps and a column for sleep stage annotations.
}
\examples{
edf <- system.file("extdata", "mini.edf", package = "nocturn")
read_edf_epochs(edf)
}
\seealso{
Other data loading:
\code{\link[=load_batch]{load_batch()}},
\code{\link[=load_epochs]{load_epochs()}},
\code{\link[=load_sessions]{load_sessions()}},
\code{\link[=read_edf_sessions]{read_edf_sessions()}}
}
\concept{data loading}
