% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{select_devices}
\alias{select_devices}
\title{Select devices by ID}
\usage{
select_devices(sessions, device_ids, return_mask = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{device_ids}{The device IDs to select}

\item{return_mask}{If TRUE, returns a logical vector indicating which sessions were recorded by the specified devices}
}
\value{
The sessions dataframe with only the sessions recorded by the specified devices, or a logical vector if \code{return_mask} is TRUE
}
\description{
Select devices by ID
}
\details{
This function uses columns:
\itemize{
\item \code{device_id}
}
}
\examples{
filtered_sessions <- select_devices(example_sessions, c("VTGVSRTHCA"))
}
\seealso{
\code{\link[=select_subjects]{select_subjects()}} to select sessions by subject ID.

Other filtering:
\code{\link[=filter_by_age_range]{filter_by_age_range()}},
\code{\link[=filter_by_night_range]{filter_by_night_range()}},
\code{\link[=filter_by_sex]{filter_by_sex()}},
\code{\link[=filter_epochs_from_sessions]{filter_epochs_from_sessions()}},
\code{\link[=keep_longest]{keep_longest()}},
\code{\link[=remove_sessions_no_sleep]{remove_sessions_no_sleep()}},
\code{\link[=select_subjects]{select_subjects()}},
\code{\link[=set_min_sleep_period]{set_min_sleep_period()}},
\code{\link[=set_min_time_in_bed]{set_min_time_in_bed()}},
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}},
\code{\link[=set_session_start_time_range]{set_session_start_time_range()}}
}
\concept{filtering}
