% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_accessors.R
\name{set_colnames}
\alias{set_colnames}
\title{Set column names for a data frame (sessions or epochs)}
\usage{
set_colnames(df, col)
}
\arguments{
\item{df}{A data frame}

\item{col}{A named list of column names}
}
\value{
The data frame with updated "col" attribute
}
\description{
This function updates the "col" attribute for a dataframe, overwriting existing column names with the ones provided.
}
\examples{
sessions <- set_colnames(
 example_sessions,
 list(night = "calendar_date", sleep_period = "duration_sleep")
)
}
\seealso{
Other internal:
\code{\link[=check_suggests]{check_suggests()}},
\code{\link[=get_col]{get_col()}},
\code{\link[=get_colnames]{get_colnames()}},
\code{\link[=get_epoch_colnames]{get_epoch_colnames()}},
\code{\link[=get_session_colnames]{get_session_colnames()}},
\code{\link[=get_time_per_day]{get_time_per_day()}},
\code{\link[=is_iso8601_datetime]{is_iso8601_datetime()}},
\code{\link[=set_data_type]{set_data_type()}}
}
\concept{internal}
