% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{set_data_type}
\alias{set_data_type}
\title{Set the type attribute for a data frame}
\usage{
set_data_type(data, type)
}
\arguments{
\item{data}{A data frame containing either sessions or epochs data}

\item{type}{A string indicating the type of data (e.g. "sessions" or "epochs")}
}
\value{
The data frame with updated "type" attribute
}
\description{
This function sets the "type" attribute for a data frame to indicate whether it contains session or epoch data.
}
\examples{
set_data_type(example_sessions, "sessions")
}
\seealso{
Other internal:
\code{\link[=check_suggests]{check_suggests()}},
\code{\link[=get_col]{get_col()}},
\code{\link[=get_colnames]{get_colnames()}},
\code{\link[=get_epoch_colnames]{get_epoch_colnames()}},
\code{\link[=get_session_colnames]{get_session_colnames()}},
\code{\link[=get_time_per_day]{get_time_per_day()}},
\code{\link[=is_iso8601_datetime]{is_iso8601_datetime()}},
\code{\link[=set_colnames]{set_colnames()}}
}
\concept{internal}
