% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{set_session_start_time_range}
\alias{set_session_start_time_range}
\title{Set session start time range}
\usage{
set_session_start_time_range(sessions, from_time, to_time, return_mask = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{from_time}{Include sessions that started after this time (in format HH:MM)}

\item{to_time}{Include sessions that started before this time (in format HH:MM)}

\item{return_mask}{If TRUE, returns a logical vector indicating which sessions meet the time range requirement}
}
\value{
The sessions dataframe with only the sessions that started within the specified time range, or a logical vector if \code{return_mask} is TRUE
}
\description{
Set session start time range
}
\details{
This function uses columns:
\itemize{
\item \code{session_start}
}
}
\examples{
filtered_sessions <- set_session_start_time_range(example_sessions, "22:00", "06:00")
}
\seealso{
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}} to filter sessions based on sleep onset time.

Other filtering:
\code{\link[=filter_by_age_range]{filter_by_age_range()}},
\code{\link[=filter_by_night_range]{filter_by_night_range()}},
\code{\link[=filter_by_sex]{filter_by_sex()}},
\code{\link[=filter_epochs_from_sessions]{filter_epochs_from_sessions()}},
\code{\link[=remove_sessions_no_sleep]{remove_sessions_no_sleep()}},
\code{\link[=select_devices]{select_devices()}},
\code{\link[=select_subjects]{select_subjects()}},
\code{\link[=set_min_sleep_period]{set_min_sleep_period()}},
\code{\link[=set_min_time_in_bed]{set_min_time_in_bed()}},
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}}
}
\concept{filtering}
