% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{shift_times_by_12h}
\alias{shift_times_by_12h}
\title{Shift times to break at 12 pm}
\usage{
shift_times_by_12h(times)
}
\arguments{
\item{times}{A vector of times in POSIXct format, character convertible to POSIXct, or numerical (in hours).}
}
\value{
A vector of times in POSIXct format (or numerical if numerical provided as input) shifted to start at 12 PM
}
\description{
This function shifts times so that the day starts at 12 PM.
This is useful for plotting night data
}
\examples{
# Shift a vector of times in HH:MM format
shift_times_by_12h(c("02:30", "16:00"))
#> "14:30" "04:00"

# Shift times in YYYY-MM-DD HH:MM:SS format
shift_times_by_12h(c("2025-04-08 23:00:00", "2025-04-09 01:00:00"))
#> "2025-04-08 11:00" "2025-04-09 13:00"

# Shift sessions start times to start at 12 PM
shifted_times <- shift_times_by_12h(example_sessions$session_start)

# Use dplyr::mutate to dicrectly add the shifted times to a dataframe
epochs <- example_epochs |>
  dplyr::mutate(shifted_time = shift_times_by_12h(timestamp))
}
\seealso{
Other time processing:
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}},
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}},
\code{\link[=max_time]{max_time()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_date]{parse_date()}},
\code{\link[=parse_time]{parse_time()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=time_diff]{time_diff()}},
\code{\link[=time_to_hours]{time_to_hours()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
