% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_helpers.R
\name{show_log_once}
\alias{show_log_once}
\title{Show a log message only once based on a condition}
\usage{
show_log_once(logger, condition, log_shown, log_msg, log_type)
}
\arguments{
\item{logger}{The logger to write the text to}

\item{condition}{A logical condition to determine whether to log the message}

\item{log_shown}{A reactive value indicating whether the log message has been shown}

\item{log_msg}{The message to log}

\item{log_type}{One of \code{default}, \code{info}, \code{error}, \code{warning}, \code{starting}, \code{complete}}
}
\value{
No return value, called for side effects
}
\description{
This function logs a message only once when a specified condition is true.
It uses a reactive value to track whether the message has already been logged.
}
\keyword{internal}
