% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_metrics.R
\name{sleep_regularity_index}
\alias{sleep_regularity_index}
\title{Calculate the Sleep Regularity Index (SRI)}
\usage{
sleep_regularity_index(epochs)
}
\arguments{
\item{epochs}{The epochs data frame}
}
\value{
The Sleep Regularity Index (SRI) value
}
\description{
The Sleep Regularity Index (SRI) is a measure of the regularity of sleep patterns.
It is calculated as the percentage of epochs where the sleep state remains the same after 24 hours.
}
\details{
This function uses columns:
\itemize{
\item \code{timestamp}
\item \code{is_asleep}
}
}
\examples{
sleep_regularity_index(example_epochs)
}
\seealso{
Other sleep metrics:
\code{\link[=chronotype]{chronotype()}},
\code{\link[=composite_phase_deviation]{composite_phase_deviation()}},
\code{\link[=interdaily_stability]{interdaily_stability()}},
\code{\link[=social_jet_lag]{social_jet_lag()}}
}
\concept{sleep metrics}
