% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_report.R
\name{sleep_report}
\alias{sleep_report}
\title{Generate a patient sleep report in PDF format}
\usage{
sleep_report(sessions, title = "", output_file = "Sleep_report.pdf")
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{title}{The title of the report. Default is an empty string.}

\item{output_file}{Path for the output PDF. Default is "Sleep_report.pdf"}
}
\value{
No return value. Called for side-effects
}
\description{
This function generates a sleep report in PDF format using an SVG template.
It is designed to work with Somnofy data. Other data types may be supported in the future.
}
\details{
This function uses columns:
\itemize{
\item \code{night}
\item \code{time_at_sleep_onset}
\item \code{time_at_wakeup}
\item \code{time_at_midsleep}
\item \code{sleep_onset_latency}
\item \code{sleep_period}
\item \code{time_in_bed}
}
}
\examples{
\donttest{sleep_report(example_sessions)}
}
