% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_distributions.R
\name{sleeptimes_boxplot}
\alias{sleeptimes_boxplot}
\title{Plot boxplots for sleep onset, midsleep, and wakeup times}
\usage{
sleeptimes_boxplot(sessions, circular = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{circular}{Whether to output a circular plot (default FALSE)}
}
\value{
A ggplot object with three horizontal boxplots (onset, midsleep, wakeup)
}
\description{
Plot boxplots for sleep onset, midsleep, and wakeup times
}
\details{
This function uses columns:
\itemize{
\item \code{time_at_sleep_onset}
\item \code{time_at_wakeup}
\item \code{time_at_midsleep}
}
}
\examples{
sleeptimes_boxplot(example_sessions)
}
