% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_distributions.R
\name{sleeptimes_density}
\alias{sleeptimes_density}
\title{Plot density curves for sleep onset, midsleep, and wakeup times with a dashed line showing the median}
\usage{
sleeptimes_density(sessions, adjust = 1, circular = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{adjust}{The bandwidth adjustment for the density estimate (default 1)}

\item{circular}{Whether to output a circular plot (default FALSE)}
}
\value{
A ggplot object with three overlaid density curves (sleep onset, midsleep, wakeup)
}
\description{
Plot density curves for sleep onset, midsleep, and wakeup times with a dashed line showing the median
}
\details{
This function uses columns:
\itemize{
\item \code{time_at_sleep_onset}
\item \code{time_at_wakeup}
\item \code{time_at_midsleep}
}
}
\examples{
sleeptimes_density(example_sessions)
}
