% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{standardise_types}
\alias{standardise_types}
\title{Standardise column types using a list of parser functions}
\usage{
standardise_types(df, parsers)
}
\arguments{
\item{df}{The dataframe to standardise}

\item{parsers}{A list of parsing functions}
}
\description{
Standardise column types using a list of parser functions
}
\keyword{internal}
