% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{time_diff}
\alias{time_diff}
\title{Compute the forward time difference from t1 to t2 (wrapping at 24)}
\usage{
time_diff(t1, t2, unit = "hour")
}
\arguments{
\item{t1}{First time (character, POSIXct, or numeric hour)}

\item{t2}{Second time (character, POSIXct, or numeric hour)}

\item{unit}{The unit of time. Can be "second", "minute", or "hour". Default is "hour".}
}
\value{
The forward difference in the specified unit (numeric, always positive, 0 <= x < 24)
}
\description{
This function returns the time from t1 to t2, always moving forward on the clock.
For example, from 07:00 to 22:00 is 15 hours, from 22:00 to 07:00 is 9 hours.
}
\examples{
time_diff("07:00", "22:00") # 15
time_diff("22:00", "07:00") # 9
time_diff("07:00", "22:00", unit = "minute") # 540
}
\seealso{
Other time processing:
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}},
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}},
\code{\link[=max_time]{max_time()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_date]{parse_date()}},
\code{\link[=parse_time]{parse_time()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=shift_times_by_12h]{shift_times_by_12h()}},
\code{\link[=time_to_hours]{time_to_hours()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
