% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{time_to_hours}
\alias{time_to_hours}
\title{Convert time vector to numeric hours}
\usage{
time_to_hours(time_vector)
}
\arguments{
\item{time_vector}{A vector of time strings}
}
\value{
A numeric vector representing the time in hours
}
\description{
This function converts a vector of time strings or POSIXct objects to numeric hours.
}
\details{
See \code{\link[=parse_time]{parse_time()}} for supported time formats.
}
\examples{
time_to_hours(c("2026-10-12T14:20:09"))
}
\seealso{
Other time processing:
\code{\link[=group_epochs_by_night]{group_epochs_by_night()}},
\code{\link[=group_sessions_by_night]{group_sessions_by_night()}},
\code{\link[=max_time]{max_time()}},
\code{\link[=mean_time]{mean_time()}},
\code{\link[=min_time]{min_time()}},
\code{\link[=parse_date]{parse_date()}},
\code{\link[=parse_time]{parse_time()}},
\code{\link[=sd_time]{sd_time()}},
\code{\link[=shift_times_by_12h]{shift_times_by_12h()}},
\code{\link[=time_diff]{time_diff()}},
\code{\link[=update_date]{update_date()}}
}
\concept{time processing}
