% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_by_night_range}
\alias{filter_by_night_range}
\title{Filter sessions for nights within a night range}
\usage{
filter_by_night_range(sessions, from_night, to_night, return_mask = FALSE)
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{from_night}{The start night of the range (inclusive) in YYYY-MM-DD format}

\item{to_night}{The end night of the range (inclusive) in YYYY-MM-DD format}

\item{return_mask}{If TRUE, returns a logical vector indicating which sessions meet the night range requirement}
}
\value{
The sessions dataframe with only the sessions that fall within the specified night range, or a logical vector if \code{return_mask} is TRUE
}
\description{
Filter sessions for nights within a night range
}
\details{
This function uses columns:
\itemize{
\item \code{night}
}
}
\examples{
filtered_sessions <- filter_by_night_range(example_sessions, "2025-04-07", "2025-04-10")
}
\seealso{
Other filtering:
\code{\link[=filter_by_age_range]{filter_by_age_range()}},
\code{\link[=filter_by_sex]{filter_by_sex()}},
\code{\link[=filter_epochs_from_sessions]{filter_epochs_from_sessions()}},
\code{\link[=keep_longest]{keep_longest()}},
\code{\link[=remove_sessions_no_sleep]{remove_sessions_no_sleep()}},
\code{\link[=select_devices]{select_devices()}},
\code{\link[=select_subjects]{select_subjects()}},
\code{\link[=set_min_sleep_period]{set_min_sleep_period()}},
\code{\link[=set_min_time_in_bed]{set_min_time_in_bed()}},
\code{\link[=set_session_sleep_onset_range]{set_session_sleep_onset_range()}},
\code{\link[=set_session_start_time_range]{set_session_start_time_range()}}
}
\concept{filtering}
