% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\title{Create a Bland-Altman plot}
\usage{
plot_bland_altman(sessions1, sessions2, variable)
}
\arguments{
\item{sessions1}{The first sessions dataframe to be compared}

\item{sessions2}{The second sessions dataframe to be compared}

\item{variable}{The name of the variable to compare. Can be a single value for both dataframes, or an array of two values, one for each dataframe}
}
\value{
a ggplot object showing the Bland-Altman plot
}
\description{
Create a Bland-Altman plot
}
\details{
This function uses columns:
\itemize{
\item \code{night}
\item \code{sleep_period}
}
}
\examples{
plot_bland_altman(example_sessions, example_sessions, "time_at_sleep_onset")
}
\concept{comparison plot}
