% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_clock.R
\name{plot_sleep_clock}
\alias{plot_sleep_clock}
\title{Plot Sleep Clock}
\usage{
plot_sleep_clock(sessions, color_by = "default")
}
\arguments{
\item{sessions}{The sessions dataframe}

\item{color_by}{The variable to color the segments by. Can be "default" or any other column name in the sessions dataframe.}
}
\value{
A ggplot object showing the sleep clock
}
\description{
Plot Sleep Clock
}
\details{
This function uses columns:
\itemize{
\item \code{time_at_sleep_onset}
\item \code{time_at_wakeup}
\item \code{night}
}
}
\examples{
plot_sleep_clock(example_sessions)
}
\seealso{
Other plot sessions:
\code{\link[=plot_sleep_bubbles]{plot_sleep_bubbles()}},
\code{\link[=plot_timeseries_comparison]{plot_timeseries_comparison()}},
\code{\link[=plot_timeseries_sessions]{plot_timeseries_sessions()}}
}
\concept{plot sessions}
